update bank class, add default constructor.

+++ bank.cpp
@@ -6,7 +6,7 @@ Bank::Bank(QString bankName, QJsonObject bankInfo)
     isAvailable = false;
     if(!bankInfo.isEmpty()) {
         this->name = bankName;
-        this->transferFee = bankInfo.take("transferfee").toInt();
+        this->transferFee = bankInfo.take("transferfee").toDouble();
         isAvailable = true;
         return;
     }
